fix(sentry): Update sentry and use local client to avoid polluting global scope
