+++ .vscode/settings.json
@@ -1,8 +1,8 @@
 {
   "editor.codeActionsOnSave": {
-    "source.fixAll.eslint": true
+    "source.fixAll.eslint": "explicit"
   },
-  "editor.formatOnSave": false,
+  "editor.formatOnSave": true,
   "eslint.packageManager": "yarn",
   "npm.packageManager": "yarn",
   "typescript.tsc.autoDetect": "off",
