+++ package.json
@@ -197,7 +197,7 @@
   "dependencies": {},
   "devDependencies": {
     "@amplitude/node": "^1.10.2",
-    "@sentry/node": "^7.47.0",
+    "@sentry/browser": "^8.35.0",
     "@types/chai": "^4.3.4",
     "@types/chai-as-promised": "^7.1.5",
     "@types/glob": "^7.2.0",
